Add files via upload
You are playing an online game. In the game, a numbers is displayed on the screen. In order to win the game, you have to Count the trailing zeros in the factorial value of the given number. Write an algorithm to
count the trailing zeros in the factorial value of the given number.
Input
The input consists of an integer num, representing the number
displayed on the screen. Output
Print An integer representing the count of trailing zeros in the factorial
of the given numbers. Note
The factorial of the number is calculated as the product of integer
numbers from 1 to num.
 Example
Input
5
Output:
1
Explanation:
On calculating the factorial of 5, the output is 120 (1 x2x3x4x5). There is
only one trailing 0 in 120.so the output is 1