Add files via upload
A company wishes to transmit data to another server. The data consists of numbers only. To secure the data during transmission, they plan to reverse the data first. Write an algorithm to reverse the data. 
Input
The input consists of an integer data, representing the data tobetransmitted. Output
Print an integer representing the given data in reverse form. 
Example
Input
5783789
Output:
9873875
Explanation:
On reversing the given value, the output is 9873875.